Added rudimentary integration testing for Goal agent tests.
This works with running one or more test2g files through the command
line with the Goal jar. There is a issue though in Goal that gives
a Nullpointerexception where it should not be. After this is fixed
independently this will be working integration testing.

A simple regex expression of the output of the tests is used to find if
the tests failed or passed.