Test if user and email are processed when first running the connector.
This commit is a temporary solution to see if it works. The connector fork
is used is currently set to setupTravis branch. When the pull request for the integration
testing for the connector in https://github.com/levilime/tygron is merged
the CI script will be to first pull the connector from the repo, then run
the maven install and tests and from there the user and email information
should be in the right place for the Tygron goal agent.